# Improve dependency resolution performance by not clearing caches when backtracking