--- a/src/poetry/mixology/version_solver.py
+++ b/src/poetry/mixology/version_solver.py
@@ -39,15 +39,20 @@ class DependencyCache:
     """
 
     def __init__(self, provider: Provider) -> None:
-        self.provider = provider
-        self.cache: dict[
-            tuple[str, str | None, str | None, str | None, str | None],
-            list[DependencyPackage],
-        ] = {}
+        self._provider = provider
+        self._cache: dict[
+            int,
+            dict[
+                tuple[str, str | None, str | None, str | None, str | None],
+                list[DependencyPackage],
+            ],
+        ] = collections.defaultdict(dict)
 
         self.search_for = functools.lru_cache(maxsize=128)(self._search_for)
 
-    def _search_for(self, dependency: Dependency) -> list[DependencyPackage]:
+    def _search_for(
+        self, dependency: Dependency, level: int
+    ) -> list[DependencyPackage]:
         key = (
             dependency.complete_name,
             dependency.source_type,
@@ -56,12 +61,17 @@ class DependencyCache:
             dependency.source_subdirectory,
         )
 
-        packages = self.cache.get(key)
-
-        if packages:
-            packages = [
-                p for p in packages if dependency.constraint.allows(p.package.version)
-            ]
+        for check_level in range(level, -1, -1):
+            packages = self._cache[check_level].get(key)
+            if packages is not None:
+                packages = [
+                    p
+                    for p in packages
+                    if dependency.constraint.allows(p.package.version)
+                ]
+                break
+        else:
+            packages = None
 
         # provider.search_for() normally does not include pre-release packages
         # (unless requested), but will include them if there are no other
@@ -71,14 +81,14 @@ class DependencyCache:
         # nothing, we need to call provider.search_for() again as it may return
         # additional results this time.
         if not packages:
-            packages = self.provider.search_for(dependency)
+            packages = self._provider.search_for(dependency)
 
-        self.cache[key] = packages
-
+        self._cache[level][key] = packages
         return packages
 
-    def clear(self) -> None:
-        self.cache.clear()
+    def clear_level(self, level: int) -> None:
+        self.search_for.cache_clear()
+        self._cache.pop(level, None)
 
 
 class VersionSolver:
@@ -318,9 +328,9 @@ class VersionSolver:
                     self._solution.decision_level, previous_satisfier_level, -1
                 ):
                     self._contradicted_incompatibilities.pop(level, None)
+                    self._dependency_cache.clear_level(level)
 
                 self._solution.backtrack(previous_satisfier_level)
-                self._dependency_cache.clear()
                 if new_incompatibility:
                     self._add_incompatibility(incompatibility)
 
@@ -418,7 +428,11 @@ class VersionSolver:
                 if locked:
                     return is_specific_marker, Preference.LOCKED, 1
 
-            num_packages = len(self._dependency_cache.search_for(dependency))
+            num_packages = len(
+                self._dependency_cache.search_for(
+                    dependency, self._solution.decision_level
+                )
+            )
 
             if num_packages < 2:
                 preference = Preference.NO_CHOICE
@@ -435,7 +449,9 @@ class VersionSolver:
 
         locked = self._provider.get_locked(dependency)
         if locked is None:
-            packages = self._dependency_cache.search_for(dependency)
+            packages = self._dependency_cache.search_for(
+                dependency, self._solution.decision_level
+            )
             package = next(iter(packages), None)
 
             if package is None: