--- a/tests/mixology/version_solver/test_dependency_cache.py
+++ b/tests/mixology/version_solver/test_dependency_cache.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from copy import deepcopy
 from typing import TYPE_CHECKING
+from unittest import mock
 
 from poetry.factory import Factory
 from poetry.mixology.version_solver import DependencyCache
@@ -32,14 +33,14 @@ def test_solver_dependency_cache_respects_source_type(
     cache.search_for.cache_clear()
 
     # ensure cache was never hit for both calls
-    cache.search_for(dependency_pypi)
-    cache.search_for(dependency_git)
+    cache.search_for(dependency_pypi, 0)
+    cache.search_for(dependency_git, 0)
     assert not cache.search_for.cache_info().hits
 
     # increase test coverage by searching for copies
     # (when searching for the exact same object, __eq__ is never called)
-    packages_pypi = cache.search_for(deepcopy(dependency_pypi))
-    packages_git = cache.search_for(deepcopy(dependency_git))
+    packages_pypi = cache.search_for(deepcopy(dependency_pypi), 0)
+    packages_git = cache.search_for(deepcopy(dependency_git), 0)
 
     assert cache.search_for.cache_info().hits == 2
     assert cache.search_for.cache_info().currsize == 2
@@ -58,6 +59,44 @@ def test_solver_dependency_cache_respects_source_type(
     assert package_git.package.source_type == dependency_git.source_type
     assert package_git.package.source_url == dependency_git.source_url
     assert package_git.package.source_resolved_reference == MOCK_DEFAULT_GIT_REVISION
+
+
+def test_solver_dependency_cache_pulls_from_prior_level_cache(
+    root: ProjectPackage, provider: Provider, repo: Repository
+) -> None:
+    dependency_pypi = Factory.create_dependency("demo", ">=0.1.0")
+    root.add_dependency(dependency_pypi)
+    add_to_repo(repo, "demo", "1.0.0")
+
+    wrapped_provider = mock.Mock(wraps=provider)
+    cache = DependencyCache(wrapped_provider)
+    cache.search_for.cache_clear()
+
+    # On first call, provider.search_for() should be called and the level-0
+    # cache populated.
+    cache.search_for(dependency_pypi, 0)
+    assert len(wrapped_provider.search_for.mock_calls) == 1
+    assert ("demo", None, None, None, None) in cache._cache[0]
+    assert cache.search_for.cache_info().hits == 0
+    assert cache.search_for.cache_info().misses == 1
+
+    # On second call at level 1, provider.search_for() should not be called
+    # again and the level-1 cache should be populated from the level-0 cache.
+    cache.search_for(dependency_pypi, 1)
+    assert len(wrapped_provider.search_for.mock_calls) == 1
+    assert ("demo", None, None, None, None) in cache._cache[1]
+    assert cache._cache[0] == cache._cache[1]
+    assert cache.search_for.cache_info().hits == 0
+    assert cache.search_for.cache_info().misses == 2
+
+    # Clearing the level 1 cache should invalidate the lru_cache on
+    # cache.search_for and wipe out the level 1 cache while preserving the
+    # level 0 cache.
+    cache.clear_level(1)
+    assert set(cache._cache.keys()) == {0}
+    assert ("demo", None, None, None, None) in cache._cache[0]
+    assert cache.search_for.cache_info().hits == 0
+    assert cache.search_for.cache_info().misses == 0
 
 
 def test_solver_dependency_cache_respects_subdirectories(
@@ -87,14 +126,14 @@ def test_solver_dependency_cache_respects_subdirectories(
     cache.search_for.cache_clear()
 
     # ensure cache was never hit for both calls
-    cache.search_for(dependency_one)
-    cache.search_for(dependency_one_copy)
+    cache.search_for(dependency_one, 0)
+    cache.search_for(dependency_one_copy, 0)
     assert not cache.search_for.cache_info().hits
 
     # increase test coverage by searching for copies
     # (when searching for the exact same object, __eq__ is never called)
-    packages_one = cache.search_for(deepcopy(dependency_one))
-    packages_one_copy = cache.search_for(deepcopy(dependency_one_copy))
+    packages_one = cache.search_for(deepcopy(dependency_one), 0)
+    packages_one_copy = cache.search_for(deepcopy(dependency_one_copy), 0)
 
     assert cache.search_for.cache_info().hits == 2
     assert cache.search_for.cache_info().currsize == 2