--- a/tests/mixology/version_solver/test_dependency_cache.py
+++ b/tests/mixology/version_solver/test_dependency_cache.py
@@ -30,20 +30,20 @@ def test_solver_dependency_cache_respects_source_type(
     add_to_repo(repo, "demo", "1.0.0")
 
     cache = DependencyCache(provider)
-    cache.search_for.cache_clear()
+    cache._search_for_cached.cache_clear()
 
     # ensure cache was never hit for both calls
     cache.search_for(dependency_pypi, 0)
     cache.search_for(dependency_git, 0)
-    assert not cache.search_for.cache_info().hits
+    assert not cache._search_for_cached.cache_info().hits
 
     # increase test coverage by searching for copies
     # (when searching for the exact same object, __eq__ is never called)
     packages_pypi = cache.search_for(deepcopy(dependency_pypi), 0)
     packages_git = cache.search_for(deepcopy(dependency_git), 0)
 
-    assert cache.search_for.cache_info().hits == 2
-    assert cache.search_for.cache_info().currsize == 2
+    assert cache._search_for_cached.cache_info().hits == 2
+    assert cache._search_for_cached.cache_info().currsize == 2
 
     assert len(packages_pypi) == len(packages_git) == 1
     assert packages_pypi != packages_git
@@ -65,38 +65,59 @@ def test_solver_dependency_cache_pulls_from_prior_level_cache(
     root: ProjectPackage, provider: Provider, repo: Repository
 ) -> None:
     dependency_pypi = Factory.create_dependency("demo", ">=0.1.0")
+    dependency_pypi_constrained = Factory.create_dependency("demo", ">=0.1.0,<2.0.0")
     root.add_dependency(dependency_pypi)
+    root.add_dependency(dependency_pypi_constrained)
     add_to_repo(repo, "demo", "1.0.0")
 
     wrapped_provider = mock.Mock(wraps=provider)
     cache = DependencyCache(wrapped_provider)
-    cache.search_for.cache_clear()
+    cache._search_for_cached.cache_clear()
 
-    # On first call, provider.search_for() should be called and the level-0
-    # cache populated.
+    # On first call, provider.search_for() should be called and the cache
+    # populated.
     cache.search_for(dependency_pypi, 0)
     assert len(wrapped_provider.search_for.mock_calls) == 1
-    assert ("demo", None, None, None, None) in cache._cache[0]
-    assert cache.search_for.cache_info().hits == 0
-    assert cache.search_for.cache_info().misses == 1
+    assert ("demo", None, None, None, None) in cache._cache
+    assert ("demo", None, None, None, None) in cache._cached_dependencies_by_level[0]
+    assert cache._search_for_cached.cache_info().hits == 0
+    assert cache._search_for_cached.cache_info().misses == 1
 
-    # On second call at level 1, provider.search_for() should not be called
-    # again and the level-1 cache should be populated from the level-0 cache.
+    # On second call at level 1, neither provider.search_for() nor
+    # cache._search_for_cached() should have been called again, and the cache
+    # should remain the same.
     cache.search_for(dependency_pypi, 1)
     assert len(wrapped_provider.search_for.mock_calls) == 1
-    assert ("demo", None, None, None, None) in cache._cache[1]
-    assert cache._cache[0] == cache._cache[1]
-    assert cache.search_for.cache_info().hits == 0
-    assert cache.search_for.cache_info().misses == 2
+    assert ("demo", None, None, None, None) in cache._cache
+    assert ("demo", None, None, None, None) in cache._cached_dependencies_by_level[0]
+    assert set(cache._cached_dependencies_by_level.keys()) == {0}
+    assert cache._search_for_cached.cache_info().hits == 1
+    assert cache._search_for_cached.cache_info().misses == 1
 
-    # Clearing the level 1 cache should invalidate the lru_cache on
-    # cache.search_for and wipe out the level 1 cache while preserving the
+    # On third call at level 2 with an updated constraint for the `demo`
+    # package should not call provider.search_for(), but should call
+    # cache._search_for_cached() and update the cache.
+    cache.search_for(dependency_pypi_constrained, 2)
+    assert len(wrapped_provider.search_for.mock_calls) == 1
+    assert ("demo", None, None, None, None) in cache._cache
+    assert ("demo", None, None, None, None) in cache._cached_dependencies_by_level[0]
+    assert ("demo", None, None, None, None) in cache._cached_dependencies_by_level[2]
+    assert set(cache._cached_dependencies_by_level.keys()) == {0, 2}
+    assert cache._search_for_cached.cache_info().hits == 1
+    assert cache._search_for_cached.cache_info().misses == 2
+
+    # Clearing the level 2 and level 1 caches should invalidate the lru_cache
+    # on cache.search_for and wipe out the level 2 cache while preserving the
     # level 0 cache.
+    cache.clear_level(2)
     cache.clear_level(1)
-    assert set(cache._cache.keys()) == {0}
-    assert ("demo", None, None, None, None) in cache._cache[0]
-    assert cache.search_for.cache_info().hits == 0
-    assert cache.search_for.cache_info().misses == 0
+    cache.search_for(dependency_pypi, 0)
+    assert len(wrapped_provider.search_for.mock_calls) == 1
+    assert ("demo", None, None, None, None) in cache._cache
+    assert ("demo", None, None, None, None) in cache._cached_dependencies_by_level[0]
+    assert set(cache._cached_dependencies_by_level.keys()) == {0}
+    assert cache._search_for_cached.cache_info().hits == 0
+    assert cache._search_for_cached.cache_info().misses == 1
 
 
 def test_solver_dependency_cache_respects_subdirectories(
@@ -123,20 +144,20 @@ def test_solver_dependency_cache_respects_subdirectories(
     root.add_dependency(dependency_one_copy)
 
     cache = DependencyCache(provider)
-    cache.search_for.cache_clear()
+    cache._search_for_cached.cache_clear()
 
     # ensure cache was never hit for both calls
     cache.search_for(dependency_one, 0)
     cache.search_for(dependency_one_copy, 0)
-    assert not cache.search_for.cache_info().hits
+    assert not cache._search_for_cached.cache_info().hits
 
     # increase test coverage by searching for copies
     # (when searching for the exact same object, __eq__ is never called)
     packages_one = cache.search_for(deepcopy(dependency_one), 0)
     packages_one_copy = cache.search_for(deepcopy(dependency_one_copy), 0)
 
-    assert cache.search_for.cache_info().hits == 2
-    assert cache.search_for.cache_info().currsize == 2
+    assert cache._search_for_cached.cache_info().hits == 2
+    assert cache._search_for_cached.cache_info().currsize == 2
 
     assert len(packages_one) == len(packages_one_copy) == 1