This PR improves dependency resolution performance for conflicts involving packages with many versions. It works by preserving parts of the `DependencyCache` and the `contradicted_incompatibilities` caches when backtracking.

On my machine with fully populated caches, the performance of `poetry update urllib3` (in a repo where this causes conflicts with `boto3`/`botocore`) goes from ~58 minutes to ~3 minutes with these changes.

## Background

I've been investigating reports of slow dependency resolution after urllib3 2.x was released, similar to https://github.com/python-poetry/poetry/issues/7856#issuecomment-1530011683. On my machine it takes almost an hour for Poetry to find a solution even with populated caches. I noticed that the time spent evaluating each solution increases a lot (evaluating the 10th solution takes 20ms/solution; by the 1000th solution it's taking 7000ms/solution).

After a bunch of profiling, I found two possible improvements around preserving these two caches. Currently both of these caches are cleared when backtracking which is causing a huge amount of duplicate work (almost all of the time is spent in `VersionRange.allows` according to flameprof).

Some reading of [the algorithm docs](https://github.com/dart-lang/pub/blob/master/doc/solver.md) suggests to me that we could improve this a lot by reverting the cache states back to the state of the decision level we backtrack to rather than wiping out the cache together. I updated `DependencyCache` and the `contradicted_incompatibilities` to track what decision level entries are added in and delete only the rolled back levels when backtracking.

The two cache updates are independent changes; I made them in two separate commits so that they can be merged or tested individually if desired.

## Performance improvements

All changes were tested using [this reproduction repo](https://github.com/chriskuehl/poetry-urllib3-dep-resolution-repro) and a warm cache using `poetry install && poetry update -vv urllib3`. All resulted in the same final solution after 1116 solutions evaluated.

| Branch | Seconds in version solving |
| ------ | -------------------------- |
| `master` | 3482                       |
| `DependencyCache` only | 2688 |
| `contradicted_incompatibilities` cache only | 962 |
| Both caches | 192 |

Here is a graph of time taken per solution evaluated:
![Graph of time taken per solution evaluated](https://github.com/python-poetry/poetry/assets/665269/079e522f-ab7d-4fc8-aee8-177631d9489a)

You can see that time taken to evaluate each solution still increases, but much more slowly (~200ms at solution 1115 compared to ~9000ms on master).


## Does this actually work?

I think it does, but I am not 100% sure and could use help from someone with a better understanding of the algorithm to confirm.

To try to verify, I implemented a test mode which runs both the old and new cache behavior and asserts the results of each cache operation are identical: https://github.com/chriskuehl/poetry/commits/dependency-resolution-perf-with-tests. I tested this in a bunch of repos and randomly added/updated/removed big dependency trees with only one difference discovered which I'll describe below. From the linked branch you can set `VERIFY_INCOMPATIBILITY_CACHE=1` and `VERIFY_DEPENDENCY_CACHE=1` when running Poetry operations to enable the assertions.

### The one difference from status quo I found

I found only one difference when asserting the results from the two caches: `DependencyCache.search_for` can return pre-release versions (for dependencies which do not allow them) on cache misses, but does not return them on cache hits. This happens because (even on master) the upper branch here does not filter out prerelease packages (it filters by the version range constraint only), while the bottom branch filters them out:
https://github.com/python-poetry/poetry/blob/3602b21792a1d19bddb8432ab01748419341a3ee/src/poetry/mixology/version_solver.py#L58-L64

I don't think this is a problem since they will be discarded later, but it does make my assertions sometimes fail as the caches return a different result now that the updated cache has a much higher hit rate. If I temporarily apply this patch then all assertion errors go away:

```diff
diff --git a/src/poetry/puzzle/provider.py b/src/poetry/puzzle/provider.py
index 3c3b738b..ba91c420 100644
--- a/src/poetry/puzzle/provider.py
+++ b/src/poetry/puzzle/provider.py
@@ -289,7 +289,10 @@ class Provider:
             return PackageCollection(dependency, packages)
 
         packages = self._pool.find_packages(dependency)
+        packages = [
+            p for p in packages
+            if not p.is_prerelease() or dependency.allows_prereleases()
+        ]
         packages.sort(
             key=lambda p: (
                 not p.yanked,
```

This makes me think it is not an issue? Note that I did not add this patch to this branch and instead kept the status quo behavior.