@dimbleby 2023-05-19
speculating wildly: my first worry on looking at this is the linear scan through the many caches.  Perhaps that's what the problem is in the examples where things get worse