@radoering 2023-05-19
### Regarding the difference with pre-releases

We allow pre-releases for a constraint that can only be satisfied by a pre-release even if pre-releases were not allowed explicity, i.e. `dependency.allows_prereleases()` returns `False`.

> This happens because (even on master) the upper branch here does not filter out prerelease packages (it filters by the version range constraint only), while the bottom branch filters them out:

The upper branch uses `RepositoryPool.find_packages()` which uses `Repository.find_packages()`, which only returns pre-releases if they are allowed or there is no non-pre-release that fulfills the constraint:

https://github.com/python-poetry/poetry/blob/277fa60742ec142ee965fec39363791848b6445b/src/poetry/repositories/repository.py#L46-L61

I don't think the bottom branch filters them out per se, but only if they are not allowed by the constraint. The branch uses `dependency.constraint.allows()` to filter packages, which is also used in upper branch via https://github.com/python-poetry/poetry/blob/277fa60742ec142ee965fec39363791848b6445b/src/poetry/repositories/repository.py#L99

> If I temporarily apply this patch then all assertion errors go away

This change may be problematic if only pre-releases can satisfy a constraint.

Maybe, you can verify the pre-release filtering with an example constraint like `black = "<22"`. (This constraint can only be satisfied by a pre-release of black.)