@chriskuehl 2023-05-19
> Wow, this looks quite promising. However, in some cases, it seems dependency resolution takes significantly longer now.

I've made two changes which should improve things and pushed these as separate commits. The main difference seemed to actually be the `lru_cache` hit rate on `DependencyCache.search_for` decreasing a lot due to the addition of the `level` parameter, so I've refactored that so that `level` is not needed in the cached function. I've also refactored both caches to not do linear scanning of per-level caches, so the performance should not get worse as the decision level increases anymore.

Here are the timings on my machine for your reproduction (testing with `rm -f poetry.lock && poetry lock` and averaging 3 runs):

**Timings for repro from https://github.com/python-poetry/poetry/pull/7950#issuecomment-1554285535**
| Variant | Seconds taken |
| ---- | ---- |
| `master` | 15.3 |
| Original PR | 22.2 |
| With improvement to `contradicted_incompatiblities` cache only | 22.1 |
| With improvement to `DependencyCache` only | 15.4 |
| With both improvements | 15.2 |

For me this gets the performance back down to matching master, but if you're able to re-test with these changes (on this repro or your private repo) that would be appreciated. It could be that some combinations of CPU or IO speed cause this branch to regress timings on some machines but not others. You may want to try at least 2 runs as I noticed the first run is often 1-2 seconds slower for me.

As you can see in the table above, the `contradicted_incompatiblities` improvement doesn't make much of a difference for this scenario, but it does seem to make a small but noticeable difference with my original urllib3 conflict reproduction:

**Timings for `poetry update urllib3` repro from the PR description**
| Variant | Seconds taken |
| ---- | ---- |
| `master` | 3482 |
| Original PR | 192 |
| With improvement to `contradicted_incompatiblities` cache only | 170 |
| With improvement to `DependencyCache` only | 179 |
| With both improvements | 171 |

For now I decided to keep it due to the small above improvement, but if it causes problems it wouldn't hurt too much to remove.